# Use new, faster HTML validator and parser for notes test