--- a/package.json
+++ b/package.json
@@ -27,6 +27,7 @@
   },
   "homepage": "https://github.com/mdn/browser-compat-data#readme",
   "devDependencies": {
+    "@desertnet/html-parser": "~1.0.1",
     "ajv": "~8.11.0",
     "ajv-errors": "~3.0.0",
     "ajv-formats": "~2.1.1",
@@ -34,10 +35,8 @@
     "browser-specs": "~3.10.0",
     "chalk": "~4.1.0",
     "compare-versions": "~4.1.1",
-    "html-validate": "~6.8.0",
     "mdn-confluence": "~2.2.2",
     "mocha": "~10.0.0",
-    "node-html-parser": "~5.3.3",
     "ora": "~5.4.0",
     "prettier": "~2.6.1",
     "yargs": "~17.4.0"