This PR fixes #16162 by replacing `html-validate` (and `node-html-parser`) with `@desertnet/html-parser`, which is much, much faster than our current setup.

I tried out a couple of different validators and got the following results when testing `api/AbortController.json` (while still using `node-html-parser` for validating individual tags based on our requirements):
Nothing: 0.56s
`html-validate`: 0.83s
`fast-xml-parser`: 0.56s
`node-html-parser`: 0.55s (does not offer description of issues)
`@desertnet/html-parser`: 0.56s

After completely switching to `@desertnet/html-parser` for all steps, running `npm run lint api` took 3.70 seconds (which is slightly lower than the 3.74 seconds it took including the last parser, and the 49.42 seconds it currently takes with our current HTML validator).)